Move Changes to a new file
