+++ Source/Notifications/Changes.swift
@@ -1,0 +1,54 @@
+//
+// Wire
+// Copyright (C) 2020 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import Foundation
+
+
+struct Changes: Mergeable {
+
+    // MARK: - Properties
+
+    let changedKeys: Set<String>
+    let originalChanges: [String: NSObject?]
+
+    // MARK: - Life cycle
+
+    init(
+        changedKeys: Set<String>,
+        originalChanges: [String: NSObject?] = [:]
+    ) {
+        self.changedKeys = changedKeys
+        self.originalChanges = originalChanges
+    }
+
+    // MARK: - Methods
+
+    var hasChangeInfo: Bool {
+        return !changedKeys.isEmpty || !originalChanges.isEmpty
+    }
+
+    func merged(with other: Changes) -> Changes {
+        guard other.hasChangeInfo else { return self }
+
+        return Changes(
+            changedKeys: changedKeys.union(other.changedKeys),
+            originalChanges: originalChanges.updated(other: other.originalChanges)
+        )
+    }
+
+}
